division du main en plusieurs composants
